Add meson build system
The meson system

A generated config header that contains the installation paths,
so instead of hardcoding the paths we can use that.

Install the app data inside a nwg-launchers directory so instead of
/usr/share/nwgbar and /usr/share/nwgdmenu we have
/usr/share/nwg-launchers/nwgbar and /usr/share/nwg-launchers/nwgdmenu.
It seems neater that way, but can be reversed if it is not wanted.

You can now change the prefix with meson setup builddir -Dprefix=/usr
and everything works.

Default prefix is /usr/local by default in meson.

The default style files are still copied to the user's config folder if
they do not exist - I didn't change that.

If something doesn't behave like the above, it is a bug I made.

If you need meson and ninja
sudo pip3 install meson ninja

Compile with:
meson setup builddir
meson compile -C builddir
sudo meson install -C builddir