Lock the pid file before reading/writing.
Each time a process wants to modify the pid file, it locks it or waits
on the existing lock.

Locking the .lock file helps to indicate whether there's an instance
running, but does nothing to protect the pid file correctness.
Although the probability of data race is very small,
we should account for it still.

This commit avoids the following situation:
A locks .lock file
A begins writing to the pid file | B starts
A continues writing              | B begins reading the pid file